e2e/common_io_e2e/requirements.txt
getgauge
protobuf ~= 3.20
boto3
google-cloud-storage
